Refactor: 이벤트 리스너 메서드를 클래스로 분리.
