+++ src/main/java/com/stylemycloset/notification/event/domain/ClothAttributeChangedEvent.java
@@ -1,4 +1,4 @@
-package com.stylemycloset.notification.event;
+package com.stylemycloset.notification.event.domain;
 
 public record ClothAttributeChangedEvent(
     Long clothAttributeId,
