+++ src/main/java/com/stylemycloset/notification/event/domain/FeedCommentEvent.java
@@ -1,4 +1,4 @@
-package com.stylemycloset.notification.event;
+package com.stylemycloset.notification.event.domain;
 
 import com.stylemycloset.user.entity.User;
