+++ src/main/java/com/stylemycloset/notification/event/domain/NewClothAttributeEvent.java
@@ -1,4 +1,4 @@
-package com.stylemycloset.notification.event;
+package com.stylemycloset.notification.event.domain;
 
 public record NewClothAttributeEvent(
     Long clothAttributeId,
