+++ src/main/java/com/stylemycloset/notification/event/domain/RoleChangedEvent.java
@@ -1,9 +1,9 @@
-package com.stylemycloset.notification.event;
+package com.stylemycloset.notification.event.domain;
 
 import com.stylemycloset.user.entity.Role;
 import com.stylemycloset.user.entity.User;
 
-public record RoleChangedEvent(
+public record RoleChangedEvent (
     User user,
     Role changedRole
 ) {
