+++ src/main/java/com/stylemycloset/notification/event/listener/RoleChangedNotificationEventListener.java
@@ -1,0 +1,39 @@
+package com.stylemycloset.notification.event.listener;
+
+import com.stylemycloset.notification.dto.NotificationDto;
+import com.stylemycloset.notification.entity.NotificationLevel;
+import com.stylemycloset.notification.event.domain.RoleChangedEvent;
+import com.stylemycloset.notification.service.NotificationService;
+import com.stylemycloset.sse.service.SseService;
+import com.stylemycloset.user.entity.User;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
+
+@Component
+@Slf4j
+@RequiredArgsConstructor
+public class RoleChangedNotificationEventListener {
+
+  private final NotificationService notificationService;
+  private final SseService sseService;
+
+  private static final String ROLE_CHANGED = "내 권한이 변경되었어요.";
+  private static final String ROLE_CHANGED_CONTENT = "내 권한이 [%s]에서 [%s]로 변경되었어요.";
+
+  @Async("eventTaskExecutor")
+  @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+  public void handler(RoleChangedEvent event) {
+    User user = event.user();
+    log.info("사용자 권한 변경 이벤트 호출 - UserId={}, updatedRole={} ", user.getId(), user.getRole());
+
+    String content = String.format(ROLE_CHANGED_CONTENT, user.getRole(), event.changedRole());
+    NotificationDto notificationDto = notificationService.create(user, ROLE_CHANGED, content, NotificationLevel.INFO);
+
+    sseService.sendNotification(notificationDto);
+    log.info("사용자 권한 변경 이벤트 완료 - notificationId={}", notificationDto.id());
+  }
+}
